hw-5: рефакторинг hw-4: вынос логики скролл-прогресс-бара в хук
